refactor: rename AI service files for clarity
- Rename client.ts → gemini-ai-provider.server.ts
- Rename image-to-exercise.ts → simple-text-question.ts
- Rename image-processor.ts → image-optimizer-service.ts
- Update all import references across codebase
