+++ src/lib/ai/gemini-ai-provider.server.ts
