+++ src/lib/ai/prompts/simple-text-question.ts
