+++ src/lib/ai/services/image-optimizer-service.ts
